+++ src/util/Grid.java
@@ -4,7 +4,7 @@ public class Grid extends TGrid<String, Tile> {
     public Grid(String[] lines) {
         super(lines, new TileFactory<>() {
             @Override
-            public Tile create(int row, int col, String val) {
+            public Tile create(long row, long col, String val) {
                 return new Tile(row, col, val);
             }
         });
