+++ src/util/TileFactory.java
@@ -1,5 +1,5 @@
 package util;
 
 public abstract class TileFactory<T> {
-    public abstract T create(int row, int col, String val);
+    public abstract T create(long row, long col, String val);
 }
